requirements.txt
Flask==2.3.1
python-dotenv
Flask-SQLAlchemy==3.0.3
psycopg2==2.9.6
pyjwt==2.6.0
Flask-Migrate==4.0.4
